@@ -1,10 +1,10 @@
 {
     "require-dev": {
-        "symfony/phpunit-bridge": "6.1.3",
-        "vimeo/psalm": "4.25.0",
-        "psalm/plugin-phpunit": "0.17.0",
-        "weirdan/doctrine-psalm-plugin": "2.3.0",
-        "phpunit/phpunit": "9.5.21"
+        "symfony/phpunit-bridge": "7.1.4",
+        "vimeo/psalm": "5.26.1",
+        "psalm/plugin-phpunit": "0.19.0",
+        "weirdan/doctrine-psalm-plugin": "2.9.0",
+        "phpunit/phpunit": "9.6.20"
     },
     "config": {
         "allow-plugins": {
