@@ -31,9 +31,7 @@ final class ConfiguredCountableFeature extends AbstractFeature implements HasCon
     /** @var array $packs */
     private $packs;
 
-    /** @var string $refreshPeriod */
-    private $refreshPeriod;
-
+    private string $refreshPeriod;
     private string $taxName;
     private float $taxRate;
 
@@ -56,7 +54,7 @@ final class ConfiguredCountableFeature extends AbstractFeature implements HasCon
         return $this->refreshPeriod;
     }
 
-    public function setPacks(array $packs, string $class = null): HasConfiguredPacksInterface
+    public function setPacks(array $packs, ?string $class = null): HasConfiguredPacksInterface
     {
         return $this->setPacksProperty($packs, ConfiguredCountableFeaturePack::class);
     }
@@ -64,7 +62,7 @@ final class ConfiguredCountableFeature extends AbstractFeature implements HasCon
     public function setSubscription(SubscriptionInterface $subscription): self
     {
         // If there are packs, set subscription in them, too
-        if (false === empty($this->packs)) {
+        if ([] !== $this->packs) {
             /** @var ConfiguredCountableFeaturePack $pack */
             foreach ($this->packs as $pack) {
                 $pack->setSubscription($subscription);
